+++ packages/wallet/src/wallet.ts
@@ -187,6 +187,12 @@ export class Wallet<
       case BigInt(ChainId.SKALE_NEBULA):
         gasLimit = 10000000n
         break
+      case BigInt(ChainId.SOMNIA_TESTNET):
+        gasLimit = 10000000n
+        break
+      case BigInt(ChainId.SOMNIA):
+        gasLimit = 10000000n
+        break
     }
 
     const bundle = Wallet.buildDeployTransaction(this.context, imageHash, gasLimit)